+++ tests/e2e/utils/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Added
 
 - Factories for variable product, variation, and grouped product
+- Added new constant for WordPress update page `WP_ADMIN_WP_UPDATES`
+- Added new merchant flow for `openWordPressUpdatesPage()`
+- Added new merchant flows:
+  - `openWordPressUpdatesPage()`
+  - `installAllUpdates()`
 
 # 0.1.5